requirements.txt
simplecrypto
urlparser
requests
flask
pyyaml
atlassian-python-api
Flask-Migrate